simplify package collections model
motivation: simpler is nicer

changes:
* dedupe collection.package and package, and collection.package.version and pacakge.version as they ended up very similar
* add keywords to package and package basic metadata
* adjust code & tests
