+++ Sources/PackageCollections/CMakeLists.txt
@@ -17,8 +17,10 @@ add_library(PackageCollections
   Providers/JSONPackageCollectionProvider.swift
   Providers/PackageCollectionProvider.swift
   Providers/PackageMetadataProvider.swift
+  Storage/FilePackageCollectionsSourcesStorage.swift
   Storage/PackageCollectionsSourcesStorage.swift
   Storage/PackageCollectionsStorage.swift
+  Storage/SQLitePackageCollectionsStorage.swift
   API.swift
   PackageCollections.swift
   PackageCollections+Configuration.swift
