+++ Sources/PackageCollections/Model/Collection.swift
@@ -141,74 +141,3 @@ extension PackageCollectionsModel.CollectionIdentifier: Codable {
         }
     }
 }
-
-// FIXME: add minimumPlatformVersions
-extension PackageCollectionsModel.Collection {
-    /// A representation of package metadata
-    public struct Package: Equatable, Codable {
-        public typealias Version = PackageVersion
-
-        /// Package reference
-        public let reference: PackageReference
-
-        /// Package's repository address
-        public let repository: RepositorySpecifier
-
-        /// A summary about the package
-        public let summary: String?
-
-        /// Published versions of the package
-        public let versions: [Version]
-
-        /// URL of the package's README
-        public let readmeURL: URL?
-
-        /// Initializes a `Package`
-        init(
-            repository: RepositorySpecifier,
-            summary: String?,
-            versions: [Version],
-            readmeURL: URL?
-        ) {
-            self.reference = .init(repository: repository)
-            self.repository = repository
-            self.summary = summary
-            self.versions = versions
-            self.readmeURL = readmeURL
-        }
-    }
-}
-
-extension PackageCollectionsModel.Collection {
-    /// A representation of package version
-    public struct PackageVersion: Equatable, Codable {
-        public typealias Target = PackageCollectionsModel.PackageTarget
-        public typealias Product = PackageCollectionsModel.PackageProduct
-
-        /// The version
-        public let version: TSCUtility.Version
-
-        /// The package name
-        public let packageName: String
-
-        // Custom instead of `PackageModel.Target` because we don't need the additional details
-        /// The package version's targets
-        public let targets: [Target]
-
-        // Custom instead of `PackageModel.Product` because of the simplified `Target`
-        /// The package version's products
-        public let products: [Product]
-
-        /// The package version's Swift tools version
-        public let toolsVersion: ToolsVersion
-
-        /// The package version's supported platforms verified to work
-        public let verifiedPlatforms: [PackageModel.Platform]?
-
-        /// The package version's Swift versions verified to work
-        public let verifiedSwiftVersions: [SwiftLanguageVersion]?
-
-        /// The package version's license
-        public let license: PackageCollectionsModel.License?
-    }
-}
