+++ Tests/PackageCollectionsTests/GitHubPackageMetadataProviderTests.swift
@@ -143,7 +143,7 @@ class GitHubPackageMetadataProviderTests: XCTestCase {
             let provider = GitHubPackageMetadataProvider()
             let reference = PackageReference(repository: RepositorySpecifier(url: UUID().uuidString))
             XCTAssertThrowsError(try tsc_await { callback in provider.get(reference, callback: callback) }, "should throw error") { error in
-                XCTAssertEqual(error as? GitHubPackageMetadataProvider.Errors, .unprocessable(reference))
+                XCTAssertEqual(error as? GitHubPackageMetadataProvider.Errors, .invalidGitUrl(reference.path))
             }
         }
     }
@@ -151,9 +151,9 @@ class GitHubPackageMetadataProviderTests: XCTestCase {
     func testInvalidRef() throws {
         fixture(name: "Collections") { _ in
             let provider = GitHubPackageMetadataProvider()
-            let reference = PackageReference(identity: .init(path: AbsolutePath("/")), path: "/")
+            let reference = PackageReference(identity: .init(path: AbsolutePath("/")), path: "/", kind: .local)
             XCTAssertThrowsError(try tsc_await { callback in provider.get(reference, callback: callback) }, "should throw error") { error in
-                XCTAssertEqual(error as? GitHubPackageMetadataProvider.Errors, .unprocessable(reference))
+                XCTAssertEqual(error as? GitHubPackageMetadataProvider.Errors, .invalidReferenceType(reference))
             }
         }
     }
